added important api endpoints

+++ todo_DRF/api/views.py
@@ -1,9 +1,7 @@
-from django.shortcuts import render
-from rest_framework.views import APIView
 from rest_framework.response import Response
-from rest_framework import status
 from .serializers import TaskSerializer
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
+from rest_framework.permissions import IsAuthenticated
 from .models import Task
 from rest_framework.exceptions import NotFound
 
@@ -24,22 +22,33 @@ def apiOverview(request):
 
 
 @api_view(["GET"])
+@permission_classes([IsAuthenticated])
 def taskList(request):
-    tasks = Task.objects.all()
+    tasks = Task.objects.filter(user=request.user)
     serializer = TaskSerializer(tasks, many=True)
     return Response(serializer.data)
 
 
 @api_view(["GET"])
+@permission_classes([IsAuthenticated])
 def taskDetail(request, pk):
-    tasks = Task.objects.get(id=pk)
-    serializer = TaskSerializer(tasks, many=False)
+    try:
+        task = Task.objects.get(id=pk, user=request.user)
+    except Task.DoesNotExist:
+        raise NotFound(detail="Task not found or not accessible by this user")
+
+    serializer = TaskSerializer(task, many=False)
     return Response(serializer.data)
 
 
 @api_view(["POST"])
+@permission_classes([IsAuthenticated])
 def taskUpdate(request, pk):
-    task = Task.objects.get(id=pk)
+    try:
+        task = Task.objects.get(id=pk, user=request.user)
+    except Task.DoesNotExist:
+        raise NotFound(detail="Task not found or not accessible by this user")
+
     serializer = TaskSerializer(instance=task, data=request.data)
     if serializer.is_valid():
         serializer.save()
@@ -48,8 +57,11 @@ def taskUpdate(request, pk):
 
 
 @api_view(["POST"])
+@permission_classes([IsAuthenticated])
 def taskCreate(request):
     data = request.data
+    data["user"] = request.user.id
+
     serializer = TaskSerializer(data=data)
     if serializer.is_valid():
         serializer.save()
@@ -58,11 +70,17 @@ def taskCreate(request):
 
 
 @api_view(["DELETE"])
+@permission_classes([IsAuthenticated])
 def taskDelete(request, pk):
     try:
         task = Task.objects.get(id=pk)
     except Task.DoesNotExist:
         raise NotFound(detail="Task not found")
+
+    if task.user != request.user:
+        return Response(
+            {"detail": "You do not have permission to delete this task."}, status=403
+        )
 
     task.delete()
